+++ src/main/java/com/mathdev/quickbite/services/UserService.java
@@ -3,7 +3,9 @@ package com.mathdev.quickbite.services;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.hibernate.TransientObjectException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.mathdev.quickbite.dto.OrderDTO;
@@ -15,64 +17,81 @@ import com.mathdev.quickbite.mapper.OrderMapper;
 import com.mathdev.quickbite.mapper.UserMapper;
 import com.mathdev.quickbite.repositories.OrderRepository;
 import com.mathdev.quickbite.repositories.UserRepository;
-
+import com.mathdev.quickbite.services.exceptions.DatabaseException;
+import com.mathdev.quickbite.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	UserRepository repo;
-	
+
 	@Autowired
 	OrderRepository orderRepository;
-	
-	//GET SERVICES
-	public List<UserDTO> findAll(){
-		List<User> users = repo.findAll();
-		
-		return users.stream().map(UserMapper::toDTO).toList();
-	}
-	
-	public UserDTO findById(Long id) {
-		User obj = repo.findById(id).orElseThrow(()->new RuntimeException("User not found!"));
-		
-		return UserMapper.toDTO(obj);
-	}
-	
-	public List<OrderDTO> getOrdersByUser(Long userId) {
-	    List<Order> orders = orderRepository.findByClientId(userId);
-	    return orders.stream()
-	                 .map(OrderMapper::toDTO)
-	                 .collect(Collectors.toList());
+
+	// GET SERVICES
+	public List<UserDTO> findAll() {
+		try {
+			List<User> users = repo.findAll();
+
+			return users.stream().map(UserMapper::toDTO).toList();
+		} catch (Exception e) {
+			throw new DatabaseException("Error retrieving users: " + e.getMessage());
+		}
 	}
 
-	//POST SERVICES
-	public UserDTO insert(UserInsertDTO dto) {
-		User obj = UserMapper.fromDTOInsert(dto);
-		obj = repo.save(obj);
-		
+	public UserDTO findById(Long id) {
+
+		User obj = repo.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found!"));
+
 		return UserMapper.toDTO(obj);
+
 	}
-	
-	//DELETE SERVICES
+
+	public List<OrderDTO> getOrdersByUser(Long userId) {
+		List<Order> orders = orderRepository.findByClientId(userId);
+		return orders.stream().map(OrderMapper::toDTO).collect(Collectors.toList());
+	}
+
+	// POST SERVICES
+	public UserDTO insert(UserInsertDTO dto) {
+		try {
+			User obj = UserMapper.fromDTOInsert(dto);
+
+			obj = repo.save(obj);
+
+			return UserMapper.toDTO(obj);
+		} catch (Exception e) {
+			throw new DatabaseException("Error saving product: " + e.getMessage());
+		}
+	}
+
+	// DELETE SERVICES
 	public void deleteUser(Long id) {
-		repo.deleteById(id);
+		if(!repo.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			repo.deleteById(id);
+		} catch (Exception e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
-	//PUT SERVICES
+
+	// PUT SERVICES
 	public UserDTO update(Long id, UserDTO newUser) {
-		User entity =  repo.getReferenceById(id);
+		User entity = repo.getReferenceById(id);
 		updateData(entity, newUser);
-		
+
 		return UserMapper.toDTO(repo.save(entity));
 	}
-	
-	//AUXILIARY METHODS
+
+	// AUXILIARY METHODS
 	private void updateData(User entity, UserDTO newUser) {
 		entity.setName(newUser.name());
 		entity.setEmail(newUser.email());
 		entity.setAddress(newUser.address());
 	}
-	
 
 }
